+++ scripts/fiddle-fork.sh
@@ -14,6 +14,7 @@
 # 03/23/2015 - Baseline Ver.
 # 04/03/2015 - (1) Added support for all fiddle types; (2) Added call to the fiddle-index script.
 # 04/15/2015 - See CHANGELOG @ 201504151810
+# 05/01/2015 - See CHANGELOG @ 201505011810
 # ---------------------------------------------------------------------------------------------------|
 clear
 thisFile=$(echo "$0" | sed 's/\.\///g')
@@ -36,6 +37,7 @@ then
     echo -e "\t\"jquery\"\tjQuery / Bootstrap Fiddle"
     echo -e "\t\"three\"\t\tThree.js / WebGl Fiddle"
     echo -e "\t\"chrome\"\t\tChrome Extension Fiddle"
+    echo -e "\t\"node\"\t\tNode.js Fiddle"
     echo ""
     echo -e "[a] - source fiddle directory name."
     echo ""